# perf: use is for singleton comparison