Use `is` instead of `==` for types. 
Don't use `type.__class__` for comparison if not necessary.

Not a big speed gain, but literally free and good practice.
Would you like me to split this up per file or so? Or take along in other MR's that hit these files?


master
````
codegen         mean runtime:  144.64ms, runs:   155.98,  126.17,  157.79,  134.59,  133.69,  136.52,  127.53,  126.41,  131.41,  216.32
methodcache     mean runtime:  131.88ms, runs:   131.22,  130.62,  128.15,  124.61,  119.62,  119.49,  126.34,  121.92,  185.33,  131.47
profile         mean runtime:  761.24ms, runs:   821.74,  702.32,  717.26,  702.75,  828.60,  787.26,  782.10,  776.56,  760.94,  732.83
```

new
```
codegen         mean runtime:  139.07ms, runs:   167.76,  127.62,  159.76,  124.24,  126.70,  126.35,  126.53,  134.09,  132.37,  165.23
methodcache     mean runtime:  130.56ms, runs:   135.22,  124.74,  139.52,  124.21,  123.66,  128.15,  125.82,  120.09,  162.86,  121.29
profile         mean runtime:  731.99ms, runs:   828.72,  704.10,  696.77,  705.83,  756.90,  697.19,  711.37,  719.69,  733.39,  765.91
```