@tinyb0t 2023-06-27
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py             5               5
tinygrad/graph.py                          2               2
tinygrad/lazy.py                          23              23
tinygrad/ops.py                            3               3
tinygrad/shape/shapetracker.py             1               1
tinygrad/tensor.py                         2               2
------------------------------------------------------------
total                                     36              36
------------------------------------------------------------
lines added in the tinygrad folder: 0
```